Export HTML and href escape functions
These can be used when implementing custom HTML renderers.
