+++ src/escape.rs
@@ -18,12 +18,12 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-//! Utility functions for HTML escaping
+//! Utility functions for HTML escaping. Only useful when building your own
+//! HTML renderer.
 
-use std::io;
+use std::fmt::{Arguments, Write as FmtWrite};
+use std::io::{self, ErrorKind, Write};
 use std::str::from_utf8;
-
-use crate::html::StrWrite;
 
 #[rustfmt::skip]
 static HREF_SAFE: [u8; 128] = [
@@ -41,7 +41,66 @@ static HEX_CHARS: &[u8] = b"0123456789ABCDEF";
 static AMP_ESCAPE: &str = "&amp;";
 static SLASH_ESCAPE: &str = "&#x27;";
 
-pub(crate) fn escape_href<W>(mut w: W, s: &str) -> io::Result<()>
+/// This wrapper exists because we can't have both a blanket implementation
+/// for all types implementing `Write` and types of the for `&mut W` where
+/// `W: StrWrite`. Since we need the latter a lot, we choose to wrap
+/// `Write` types.
+pub struct WriteWrapper<W>(pub W);
+
+/// Trait that allows writing string slices. This is basically an extension
+/// of `std::io::Write` in order to include `String`.
+pub trait StrWrite {
+    fn write_str(&mut self, s: &str) -> io::Result<()>;
+
+    fn write_fmt(&mut self, args: Arguments) -> io::Result<()>;
+}
+
+impl<W> StrWrite for WriteWrapper<W>
+where
+    W: Write,
+{
+    #[inline]
+    fn write_str(&mut self, s: &str) -> io::Result<()> {
+        self.0.write_all(s.as_bytes())
+    }
+
+    #[inline]
+    fn write_fmt(&mut self, args: Arguments) -> io::Result<()> {
+        self.0.write_fmt(args)
+    }
+}
+
+impl<'w> StrWrite for String {
+    #[inline]
+    fn write_str(&mut self, s: &str) -> io::Result<()> {
+        self.push_str(s);
+        Ok(())
+    }
+
+    #[inline]
+    fn write_fmt(&mut self, args: Arguments) -> io::Result<()> {
+        // FIXME: translate fmt error to io error?
+        FmtWrite::write_fmt(self, args).map_err(|_| ErrorKind::Other.into())
+    }
+}
+
+impl<W> StrWrite for &'_ mut W
+where
+    W: StrWrite,
+{
+    #[inline]
+    fn write_str(&mut self, s: &str) -> io::Result<()> {
+        (**self).write_str(s)
+    }
+
+    #[inline]
+    fn write_fmt(&mut self, args: Arguments) -> io::Result<()> {
+        (**self).write_fmt(args)
+    }
+}
+
+/// Writes an href to the buffer, escaping href unsafe bytes.
+pub fn escape_href<W>(mut w: W, s: &str) -> io::Result<()>
 where
     W: StrWrite,
 {
@@ -93,7 +152,7 @@ static HTML_ESCAPES: [&'static str; 5] = ["", "&quot;", "&amp;", "&lt;", "&gt;"]
 
 /// Writes the given string to the Write sink, replacing special HTML bytes
 /// (<, >, &, ") by escape sequences.
-pub(crate) fn escape_html<W: StrWrite>(w: W, s: &str) -> io::Result<()> {
+pub fn escape_html<W: StrWrite>(w: W, s: &str) -> io::Result<()> {
     #[cfg(all(target_arch = "x86_64", feature = "simd"))]
     {
         simd::escape_html(w, s)
@@ -131,7 +190,7 @@ fn escape_html_scalar<W: StrWrite>(mut w: W, s: &str) -> io::Result<()> {
 
 #[cfg(all(target_arch = "x86_64", feature = "simd"))]
 mod simd {
-    use crate::html::StrWrite;
+    use super::StrWrite;
     use std::arch::x86_64::*;
     use std::io;
     use std::mem::size_of;
